Move crud functions to helper hook to keep things seperated better
